% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{account.tree.depth}
\alias{account.tree.depth}
\title{Get list of account tree}
\usage{
account.tree.depth(names)
}
\arguments{
\item{names}{character vector containing different account names}
}
\value{
data.frame containing column account and depth
}
\description{
For a given character vector of the account names
  generate an ordered data.frame containing account names and the
  depth of the account.

For example, "Assets:Bank:Current" account has depth 3
}

