% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{alltime.price}
\alias{alltime.price}
\title{Compute all-time-running daily average price}
\usage{
alltime.price(x, ignore.days = 30)
}
\arguments{
\item{x}{a numeric vector}

\item{ignore.days}{number of first days to ignore}
}
\value{
a numeric vector of the same size
}
\description{
Compute all-time-running daily average price
}
