% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{alluvial.plot}
\alias{alluvial.plot}
\title{alluvium plots}
\usage{
alluvial.plot(data, currency, title)
}
\arguments{
\item{data}{a data.frame with "Date", "Category", "Value" (not
necessarily this name)}

\item{currency}{currency of the values (put as ylab)}

\item{title}{plot title}
}
\value{
a ggplot object
}
\description{
alluvium plots
}
