% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{complete_dataframe_missing_dates}
\alias{complete_dataframe_missing_dates}
\title{Complete data with all dates}
\usage{
complete_dataframe_missing_dates(X, dates.series, duplicated_transactions,
  FUN)
}
\arguments{
\item{X}{dataframe with columns "Date" and "Amount"}

\item{dates.series}{a vector of dates. Entries are added to X with
the dates corresponding to dates.series}

\item{duplicated_transactions}{functions with which the several
transactions in one day are resolved}

\item{FUN}{a character vector of function names to be applied to
transactions. Each function creates a column in transaction
data.frame}
}
\value{
a data.frame with "Date" and columns corresponding ot the
  FUN names. The data.frame contains a value for each date in dates.series
}
\description{
Complete data with all dates
}
