% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{compound.poisson}
\alias{compound.poisson}
\title{estimate average/variance of transaction expense per given
  period using Poisson compound process}
\usage{
compound.poisson(data, period = 365)
}
\arguments{
\item{data}{transactions data.frame. Default is one year}

\item{period}{period in days of which the average is calculated}
}
\description{
estimate average/variance of transaction expense per given
  period using Poisson compound process
}
