% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{convert.units}
\alias{convert.units}
\title{convert units one to another}
\usage{
convert.units(data, conversion = c("1kg = 1000g"))
}
\arguments{
\item{data}{data frame containing parsed volumes}

\item{conversion}{vector of containing conversion rules. E.g.:
c("1 kg = 1000g","1l = 1000ml")}
}
\description{
convert all right units to left units
}
