% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{.period_average}
\alias{.period_average}
\title{compute a rolling average}
\usage{
.period_average(x, days)
}
\arguments{
\item{x}{a numeric vector}

\item{days}{number of days in the window}
}
\value{
a numeric vector
}
\description{
compute a rolling average
}
