% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{.period_price}
\alias{.period_price}
\title{compute a rolling price}
\usage{
.period_price(x, days)
}
\arguments{
\item{x}{a numeric vector}

\item{days}{number of days in the window}
}
\value{
a numeric vector
}
\description{
compute a rolling price
}
