% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{food.prices.convert}
\alias{food.prices.convert}
\title{Convert comments in food ledger to the corresponding prices}
\usage{
food.prices.convert(food, currency)
}
\arguments{
\item{food}{dataset with food, returns by read.ledger}

\item{currency}{sometimes currency in transaction note is given}
}
\description{
Convert comments in food ledger to the corresponding prices
}

