% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{get_tags}
\alias{get_tags}
\title{get tags and corresponding indices}
\usage{
get_tags(data)
}
\arguments{
\item{data}{data.frame, output of the read.ledger}
}
\value{
a named list with tags and corresponding indices
}
\description{
process Notes column from 'ledger csv' and generate a
  list with tags and corresponding indices in the data
}
