% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{get_tuples_tags}
\alias{get_tuples_tags}
\title{compute different combinations of tags}
\usage{
get_tuples_tags(tags, n = 3, min_num = 5)
}
\arguments{
\item{tags}{names list returned by get_tags function}

\item{n}{maximum number of tuples to consider}

\item{min_num}{minimum number of entries for the tag combinations
to appear in plots}
}
\value{
named list with indices corresponding to tuples pf tags
}
\description{
compute different combinations of tags
}
