% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{lm_forecast}
\alias{lm_forecast}
\title{linear regression forecast}
\usage{
lm_forecast(x, using_last = 365, forecast_for = 365)
}
\arguments{
\item{using_last}{number of days to use in the forecast}

\item{forecast_for}{number of days to forecast}
}
\description{
linear regression forecast
}
