% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{monthly}
\alias{monthly}
\title{Calculate 30 days average of the given vector}
\usage{
monthly(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Calculate 30 days average of the given vector
}
