% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{parse_at_entries}
\alias{parse_at_entries}
\title{parse price and volumes}
\usage{
parse_at_entries(data, re = "([0-9]+[.0-9]*)[ ]?([[:alpha:]]+)")
}
\arguments{
\item{data}{data.frame}

\item{re}{regular expression to parse}
}
\description{
get the prices and volumes columns from the parsed entries
}
