% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{plot.ledger}
\alias{plot.ledger}
\title{Plot ledger data in a current device}
\usage{
\method{plot}{ledger}(X, title, FUN = cumsum, ...)
}
\arguments{
\item{X}{data. First column date, Second column amount}

\item{title}{title for the plot}

\item{FUN}{what to do with data before plotting. This must be a
function that accept a vector as an argument and outputs a
vector.}

\item{...}{arguments passed to FUN}
}
\value{
nothing
}
\description{
A low lever function that do the actual
  plot. Transforms and plots data in some format, providing some
  extra information
}
\examples{
# read expenses data
expenses <- read.ledger("^expenses: -X EUR")

# plot cumsum over 30 days expenses
plot.ledger(X=expenses[,c(1,6)],
            title="30 days average expenses",
            FUN=filter, rep(1,30),sides=1)
}

