% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{quarterly}
\alias{quarterly}
\title{Calculate quartely averages}
\usage{
quarterly(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector of the same size as vector x
}
\description{
Calculate quartely averages
}
