% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{query.plot}
\alias{query.plot}
\title{Plot accounts for a given query}
\usage{
query.plot(query, order.depth = TRUE, order.function = function(x)
  sum(abs(x)), max.num.plots, ledger.options, ledger.path = NULL, ...)
}
\arguments{
\item{query}{query string that is used in read.ledger function to
query the ledger for transactions}

\item{order.depth}{if do ordering by account depths first}

\item{order.function}{function that is calculated in order to sort
the plots for different accounts. The function must take a vector
and return a single value}

\item{max.num.plots}{maximum number of plots}

\item{ledger.options}{extra options specified during the call to
ledger}

\item{ledger.path}{path to the ledger executable}

\item{...}{extra arguments given to plot.ledger function}
}
\value{
list of ggplots
}
\description{
For a given query plot accounts given transaction
  function (see plot.ledger). All plots are done in the current
  device.
}
\details{
the plots are ordered by the depth of the ledger account
  name and by a value of the order.function (by default it
  calculates the sum of absolute values of transactions). This
  allows to order many plots in such a fashion that firstly a
  presented a high level plots with the most impact (wrt
  order.function).
}
