% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{queryplot}
\alias{queryplot}
\title{Plot accounts for a given query}
\usage{
queryplot(query, order.function = function(x) sum(abs(x)), ledger.options,
  ...)
}
\arguments{
\item{query}{query string that is used in read.ledger function to
query the ledger for transactions}

\item{order.function}{function that is calculated in order to sort
the plots for different accounts. The function must take a vector
and return a single value}

\item{ledger.options}{extra options specified during the call to
ledger}

\item{...}{extra arguments given to plot.ledger function}
}
\value{
nothing
}
\description{
For a given query plot accounts given transaction
  function (see plot.ledger). All plots are done in the current
  device.
}
\details{
the plots are ordered by the depth of the ledger account
  name and by a value of the order.function (by default it
  calculates the sum of absolute values of transactions). This
  allows to order many plots in such a fashion that firstly a
  presented a high level plots with the most impact (wrt
  order.function).
}
