% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{read.ledger}
\alias{read.ledger}
\title{Make quries to ledger}
\usage{
read.ledger(query, options = "")
}
\arguments{
\item{query}{string query that is used in the ledger call.}

\item{options}{string of options given to ledger}
}
\value{
data.frame containing the following columns
\code{"Date",NA,"Description","Category","Currency","Amount",NA,"Notes"}
}
\description{
Call a shell and execute ledger command with a given
  query.
}
\details{
The function make the following call in shell:
  \code{ledger csv <query>}

One may pass extra argument to ledger this way.
}
\examples{
expenses <- read.ledger("^expenses: -X EUR")

assets <- read.ledger("^assets: -X EUR")

}
