% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{series.plot}
\alias{series.plot}
\title{Low level function that plots the series}
\usage{
series.plot(date, series, currency, title)
}
\arguments{
\item{date}{vector of dates}

\item{series}{vector of the same length as date containing values
of the series}

\item{currency}{type of series values (used as ylab)}

\item{title}{account name (used as a plot name)}
}
\value{
nothing
}
