% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{ssa_forecast}
\alias{ssa_forecast}
\title{ssa forecast}
\usage{
ssa_forecast(x, group, forecast_for = 365)
}
\arguments{
\item{x}{a numeric vector}

\item{group}{vector indicating which components to choose}

\item{forecast_for}{number of days to forecast}
}
\description{
ssa forecast
}
