% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{yearly}
\alias{yearly}
\title{Calculate 360 days average of the given vector}
\usage{
yearly(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector of the same size as vector x
}
