% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{account.plot}
\alias{account.plot}
\title{Plot ledger data in a current device}
\usage{
account.plot(X, title, type, categorise_accounts = NULL,
  date.interval = c(Sys.Date() - 365, Sys.Date()), FUN = c("cumsum"),
  ...)
}
\arguments{
\item{X}{data. First column date, Second column amount}

\item{title}{title for the plot (account name)}

\item{type}{type of the plot to make}

\item{categorise_accounts}{if is not NULL plot alluvial instead of
series}

\item{date.interval}{dates period between which the plots should be
done (default for one year)}

\item{FUN}{what to do with data before plotting. This must be a
function that accept a vector as an argument and outputs a
vector.}

\item{...}{arguments passed to FUN}
}
\value{
list of ggplot
}
\description{
A low lever function that do the actual
  plot. Transforms and plots data in some format, providing some
  extra information
}
