% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-functions.R
\name{adjust_forecast}
\alias{adjust_forecast}
\title{forecast values adjust

adjust forecast values with known future transaction amounts}
\usage{
adjust_forecast(forecast, dates, values, FUN = "cumsum")
}
\arguments{
\item{forecast}{vector of the forecasted values}

\item{dates}{future dates of transactions}

\item{values}{values of transactions}

\item{FUN}{function to be aplied to the adjusted values}
}
\value{
vector adjusted forecast vector
}
\description{
forecast values adjust

adjust forecast values with known future transaction amounts
}
