% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{generate.price.table}
\alias{generate.price.table}
\title{Generate a latex table with food prices}
\usage{
generate.price.table(query, ofile = "food-prices.tex", ledger.options,
  ledger.path = NULL, conversion)
}
\arguments{
\item{query}{ledger query}

\item{ofile}{output filename}

\item{ledger.options}{extra options passed to ledger}

\item{ledger.path}{path to the ledger binary}

\item{conversion}{volume unit conversion rules}
}
\description{
Generate a latex table with food prices
}
