% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{parse.notes}
\alias{parse.notes}
\title{Parse comments and get prices and volumes}
\usage{
parse.notes(data, conversion = c("1kg = 1000g"))
}
\arguments{
\item{data}{read ledger data}

\item{conversion}{unit conversion rules}
}
\description{
Parse transaction notes and get a data frame containing
columns: price, price.currency, volume, volume.currency, other comments
}
