% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{query.plot}
\alias{query.plot}
\title{Plot accounts for a given query}
\usage{
query.plot(query, type = c("amount", "price", "volume", "revalued"),
  categorise_by = c("account", "tags", "alluvial"), order.depth = TRUE,
  order.function = function(x) sum(abs(x)), max.num.plots,
  ledger.options, ledger.path = NULL, conversion = c("1kg = 1000g"),
  account_depth_for_each_tag = 1, max_tags_tuples = 3,
  min_tags_entries = 5, ...)
}
\arguments{
\item{query}{query string that is used in read.ledger function to
query the ledger for transactions}

\item{type}{type of plot to make (which amount to consider)}

\item{categorise_by}{the way plot are categorised by. In case of
tags only accounts with depth 1 are considered}

\item{order.depth}{if do ordering by account depths first}

\item{order.function}{function that is calculated in order to sort
the plots for different accounts. The function must take a vector
and return a single value}

\item{max.num.plots}{maximum number of plots}

\item{ledger.options}{extra options specified during the call to
ledger}

\item{ledger.path}{path to the ledger executable}

\item{conversion}{volume unit conversion rules}

\item{account_depth_for_each_tag}{depth of accounts considered for
each tag (in case categorise_by == "tags")}

\item{max_tags_tuples}{maximum length of tags combinations}

\item{min_tags_entries}{ignore plots that have less than 5 tags
entries}

\item{...}{extra arguments given to plot.ledger function}
}
\value{
list of ggplots
}
\description{
For a given query plot accounts given transaction
  function (see plot.ledger). All plots are done in the current
  device.
}
\details{
the plots are ordered by the depth of the ledger account
  name and by a value of the order.function (by default it
  calculates the sum of absolute values of transactions). This
  allows to order many plots in such a fashion that firstly a
  presented a high level plots with the most impact (wrt
  order.function).
}
