% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-functions.R
\name{series.plot}
\alias{series.plot}
\title{Low level function that plots the series}
\usage{
series.plot(data, currency, title)
}
\arguments{
\item{data}{data.frame to be plotted}

\item{currency}{type of series values (used as ylab)}

\item{title}{account name (used as a plot name)}
}
\value{
ggplot object
}
